feat(database): 移除点歌券自动删除功能并优化相关服务
- 删除数据库中的 cardCodeAutoDeleteDays 字段
- 移除自动清理过期点歌券的定时任务插件
- 将点歌券删除服务从 cardCodeCleanupService 拆分为独立的 cardCodeDeleteService
- 更新系统设置中相关的配置选项处理逻辑
- 清理前端界面中对应的自动删除配置表单
- 移除备份恢复功能中的自动删除天数字段处理
- 更新类型定义文件中移除相应的配置项声明
