+++ server/api/admin/backup/restore-chunk.post.ts
@@ -697,7 +697,6 @@ export default defineEventHandler(async (event) => {
               'enableSubmissionRemarks',
               'enableCardCodeRequests',
               'requireCardCodeForRequests',
-              'cardCodeAutoDeleteDays',
               'enableRequestTimeLimitation',
               'requestTimeLimitation',
               'forceBlockAllRequests',
