+++ server/api/admin/backup/restore.post.ts
@@ -1129,7 +1129,6 @@ export default defineEventHandler(async (event) => {
                           'enableSubmissionRemarks',
                           'enableCardCodeRequests',
                           'requireCardCodeForRequests',
-                          'cardCodeAutoDeleteDays',
                           'hideStudentInfo',
                           'smtpEnabled',
                           'smtpHost',
