+++ server/utils/system-settings-defaults.ts
@@ -29,7 +29,6 @@ export const SYSTEM_SETTINGS_DEFAULTS = {
   // 卡密点歌相关
   enableCardCodeRequests: false,
   requireCardCodeForRequests: false,
-  cardCodeAutoDeleteDays: 0,
   enableRequestTimeLimitation: false,
   forceBlockAllRequests: false,
   smtpEnabled: false,
@@ -70,7 +69,6 @@ export const PUBLIC_SETTINGS_FIELDS = [
   'enableSubmissionRemarks',
   'enableCardCodeRequests',
   'requireCardCodeForRequests',
-  'cardCodeAutoDeleteDays',
   'enableRequestTimeLimitation',
   'forceBlockAllRequests',
   'smtpEnabled',
